requirements.txt
django~=3.0.5
channels~=2.4.0
social-auth-app-django
tweepy
onnxruntime
pillow
numpy
scipy
scikit-image
